refactore: improve genetic algorithm code

+++ src/main/java/org/harmoniapp/geneticalgorithm/GenerationListener.java
@@ -4,5 +4,5 @@ package org.harmoniapp.geneticalgorithm;
  * Listener interface for receiving updates during the genetic algorithm's execution.
  */
 public interface GenerationListener {
-    void onGenerationUpdate(int generation, double fitness);
+    void onGenerationUpdate(double generation, double fitness);
 }
